move to recording directory
